Move relative distance deepset tests
